Configure source in preferences
This is a bit ugly.

We do not use an Adw.ComboRow because it cuts off long texts in its
dropdown, even though there's plenty of screen space.

We manually link settings with the dropdown, because I do not know how
the whole binding machinery works for list models, so we just get away
with manual signal handlers.