Remove refresh-failed signal
Instead expose the promise representing the download to the caller; this
requires that the caller handles errors, as the caller is in a better
place to take action appropriate to its context than the generic error
handler.  This prepares for a refresh scheduler which reacts on certain
errors not by displaying the error to the user, but by rescheduling
another refresh.

We still keep the signal for a completed refresh though, because this
gets directly connected to the desktop background service, etc. which
relieves callers from having to take care to actually do something with
a successful download.