Improve design of source selection
Instead of a rather ugly combination of dropdown w/ action row, let's
use an expander row with a list of action rows + radio buttons.

This works better with the size constraints of adwaita widgets, and also
allows us to present all sources at once and include the website link
for each source.

Closes #13