Split metadata download from image download
Sources now only define a function to get the metadata for the current
image, including a download URL.

Then the outer download function, which the extension class now creates,
does the actual download.

This prepares for allowing sources to return multiple images and have
the refresh service select one based on a given rule, and it removes a
significant amount of duplication from the source implementations.