Reorder initialization
Setup the refresh scheduler at the very end, when everything is wired
up, because the refresh scheduler might do an immediate refresh.

Fixes an error about a missing download function which occurred when
the extension was enabled more than six hours after the last refresh.
In this case the refresh scheduler would immediately attempt to
refresh the picture of the day, but in the previous extension order
updateDownloader() hadn't been called yet.