Do not use add-determinism just now
This reverts all the previous commits, and leaves things as they were
before.

There are no binaries for add-determinism anywhere other than in the
repos, getting it to build on Fedora is cumbersome because it needs cc,
and building it on Ubuntu then caching it somehow fails to restore.

Let's not spend too much time; instead, re-visit this once Fedora 42 is
out and we upgrade to GNOME 48.